+++ controllers/contacts-controllers.js
@@ -1,15 +1,9 @@
-const Joi = require('joi');
-
 const contactsService = require("../models/contacts");
 
 const { HttpError } = require("../helpers");
+
 const { ctrlWrapper } = require("../decorators");
 
-const contactAddSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-})
 
 const getAllContacts = async (req, res) => {
     const result = await contactsService.listContacts();
@@ -26,21 +20,12 @@ const getContactById = async (req, res) => {
 }
 
 const addContact = async (req, res) => {
-    const { error } = contactAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(404, error.message);
-    }
     const result = await contactsService.addContact(req.body);
 
     res.status(201).json(result);
 }
 
 const updateContactById = async (req, res) => {
-    const { error } = contactAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
     const { contactId } = req.params;
 
     const result = await contactsService.updateContact(contactId, req.body);
