+++ decorators/index.js
@@ -1,5 +1,7 @@
 const ctrlWrapper = require("./ctrlWrapper");
+const validateBody = require("./validateBody");
 
 module.exports = {
     ctrlWrapper,
+    validateBody,
 }
