+++ models/contacts.json
@@ -64,5 +64,11 @@
     "name": "Nikhotina Tetiana2",
     "email": "nikhotina@gmail.com",
     "phone": "(096) 657-11-52"
+  },
+  {
+    "id": "9KZHWelwLi-AAJXarbJ2_",
+    "name": "Nazar2",
+    "email": "nikhotina@gmail.com",
+    "phone": "(096) 657-11-52"
   }
 ]
