+++ routes/api/contacts.js
@@ -4,15 +4,18 @@ const router = express.Router()
 
 const contactsController = require('../../controllers/contacts-controllers');
 
+const schemas = require("../../schemas/contacts-schemas");
+
+const { validateBody } = require("../../decorators");
 
 router.get('/', contactsController.getAllContacts);
 
 router.get('/:contactId', contactsController.getContactById);
 
-router.post('/', contactsController.addContact);
+router.post('/', validateBody(schemas.contactAddSchema), contactsController.addContact);
+
+router.put('/:contactId', validateBody(schemas.contactAddSchema), contactsController.updateContactById);
 
 router.delete('/:contactId', contactsController.deleteContactById);
-
-router.put('/:contactId', contactsController.updateContactById);
 
 module.exports = router
